+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This file is used to list changes made in each version of the apache2 cookbook.
 
 -   Add helpers: `lib_dir`, `libexec_dir`, `icon_dir`
 -   Add package helpers
+-   Remove all mpm_ recipes
+-   Allow user to set the mpm mode no matter what platform they're on
+-   Set the server to listen on ports 80 and 443 by default
+-   Add dev-sec/apache-baseline tests and set cookbook defaults to securre defaults
+-   Add CircleCI
+-   Cleanup test integration directory
 
 ## v5.2.1 (04-09-2018)